Reenable broken group actions in My Devices overview screen
The group actions in My Devices overview screen were broken in the following
language translations due to translation of an identifier as reported in #793 :

 * de
 * nl
 * pt

The translations of the identifier have been reverted to the English original to mitigate the problem. Final solution should be the removal of the identifier "DeviceCheckbox" from translatable strings.