add reports to meshctrl
The only format that makes sense to me is CSV, node and mesh ids in the sessions report are just too big for terminal view and JSON is inefficient (too much white-space).

The first field (group) derives from the groupby arguement and is "0" when no grouping is applicable (e.g. db reports).

resolves #3472, resolves #3509